dom/html/test/test_track.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=833386
-->
<head>
  <meta charset='utf-8'>
  <title>Test for Bug 833386 - HTMLTrackElement</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/dom/html/test/reflect.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
reflectLimitedEnumerated({
  element: document.createElement("track"),
  attribute: "kind",
  validValues: ["subtitles", "captions", "descriptions", "chapters",
                "metadata"],
  invalidValues: ["foo", "bar", "\u0000", "null", "", "subtitle", "caption",
                  "description", "chapter", "meta"],
  defaultValue: "subtitles"
});

// Default attribute
reflectBoolean({
  element: document.createElement("track"),
  attribute: "default"
});

// Label attribute
reflectString({
  element: document.createElement("track"),
  attribute: "label",
  otherValues: [ "foo", "BAR", "_FoO", "\u0000", "null", "white space" ]
});

// Source attribute
reflectURL({
  element: document.createElement("track"),
  attribute: "src",
  otherValues: ["foo", "bar", "\u0000", "null", ""]
});

// Source Language attribute
reflectString({
  element: document.createElement("track"),
  attribute: "srclang",
  otherValues: ["foo", "bar", "\u0000", "null", ""]
});

var track = document.createElement("track");
is(track.readyState, 0, "Default ready state should be 0 (NONE).");
</script>
</pre>
</body>
</html>
